+++ app/util/env.ts
@@ -105,6 +105,7 @@ interface HarmonyEnv {
   useLocalstack: boolean;
   workFailerPeriodSec: number;
   workReaperPeriodSec: number;
+  maxErrorsForJob: number;
 }
 
 // special cases