+++ app/util/variables.ts
@@ -1,6 +1,5 @@
 import { CmrCollection, CmrUmmVariable } from './cmr';
 import { RequestValidationError } from './errors';
-import * as services from '../models/services/index';
 
 export interface HarmonyVariable {
   id: string;
@@ -23,6 +22,8 @@ export interface HarmonyRelatedUrl {
 
 interface VariableInfo {
   collectionId: string;
+  shortName: string; // collection short_name
+  versionId: string; // collection version_id
   variables?: CmrUmmVariable[];
   coordinateVariables?: CmrUmmVariable[];
 }
@@ -98,31 +99,6 @@ export function getCoordinateVariables(variables: CmrUmmVariable[]): CmrUmmVaria
 }
 
 /**
- * Get a list of variables that are defined in the service configs as being available for
- * processing for the given collection. If the returned set is empty this means there are no
- * limits set as to which variables a service will process.
- * @param collection - the CMR collection 
- * @returns A Set of variable IDs obtained from service configs
- */
-export function getSupportedVariablesForCollection(
-  collection: CmrCollection,
-): Set<string> {
-  const variableIds = new Set<string>();
-  const configs = services.getServiceConfigs();
-  for (const serviceConfig of configs) {
-    const serviceCollection = serviceConfig.collections?.find(
-      (collectionConfig) => collectionConfig.id === collection.id,
-    );
-    if (serviceCollection?.variables) {
-      for (const variableId of serviceCollection?.variables) {
-        variableIds.add(variableId);
-      }
-    }
-  }
-  return variableIds;
-}
-
-/**
  * Given a list of EOSDIS collections and variables parsed from the CMR and an OGC
  * collectionId parameter return the full variables which match.
  *
@@ -153,7 +129,8 @@ export function parseVariables(
     }
     for (const collection of eosdisCollections) {
       const coordinateVariables = getCoordinateVariables(collection.variables);
-      variableInfo.push({ collectionId: collection.id, coordinateVariables });
+      variableInfo.push({ collectionId: collection.id, shortName: collection.short_name,
+        versionId: collection.version_id, coordinateVariables });
     }
   } else {
     // Figure out which variables belong to which collections and whether any are missing.
@@ -163,17 +140,17 @@ export function parseVariables(
       // Get the list of variables configured in services.yml for this collection. If the 
       // returned set is empty then we will ignore it, otherwise we will only add variables
       // in that set
-      const serviceVariables = getSupportedVariablesForCollection(collection);
       const coordinateVariables = getCoordinateVariables(collection.variables);
       const variables = [];
       for (const variableId of variableIds) {
         const variable = collection.variables.find((v) => doesPathMatch(v, variableId));
-        if (variable && (serviceVariables.has(variableId) || serviceVariables.size === 0)) {
+        if (variable) {
           missingVariables.delete(variableId);
           variables.push(variable);
         }
       }
-      variableInfo.push({ collectionId: collection.id, variables, coordinateVariables });
+      variableInfo.push({ collectionId: collection.id, shortName: collection.short_name,
+        versionId: collection.version_id, variables, coordinateVariables });
     }
     if (missingVariables.size > 0) {
       throw new RequestValidationError(`Coverages were not found for the provided CMR collection: ${Array.from(missingVariables).join(', ')}`);
@@ -219,7 +196,8 @@ export function getVariablesForCollection(
   for (const collection of collections) {
     const coordinateVariables = getCoordinateVariables(collection.variables);
     const variables = variablesByCollection[collection.id];
-    variableInfo.push({ collectionId: collection.id, variables, coordinateVariables });
+    variableInfo.push({ collectionId: collection.id, shortName: collection.short_name,
+      versionId: collection.version_id, variables, coordinateVariables });
   }
   return variableInfo;
 }