+++ test/work-items/work-backends.ts
@@ -199,7 +199,7 @@ describe('Work Backends', function () {
       hookClearScrollSessionExpect();
 
       const failedWorkItemRecord = {
-        ...workItemRecord, ...{ status: WorkItemStatus.FAILED },
+        ...workItemRecord, ...{ status: WorkItemStatus.FAILED, scrollID: '-1234' },
       };
 
       hookWorkItemCreation(failedWorkItemRecord);
@@ -209,7 +209,8 @@ describe('Work Backends', function () {
         const updatedWorkItem = await getWorkItemById(db, this.workItem.id);
         expect(updatedWorkItem.status).to.equal(WorkItemStatus.FAILED);
       });
-      it('sets the job status is set to failed', async function () {
+
+      it('sets the job status to failed', async function () {
         const job = await Job.byJobID(db, this.job.jobID);
         expect(job.status).to.equal(JobStatus.FAILED);
       });
@@ -224,6 +225,7 @@ describe('Work Backends', function () {
         ...{
           status: WorkItemStatus.SUCCESSFUL,
           results: [getStacLocation({ id: workItemRecord.id, jobID: workItemRecord.jobID }, 'catalog.json')],
+          scrollID: '-1234',
         },
       };
       before(async () => {